myproject/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.2
django-filter==2.4.0
pytz==2021.1
sqlparse==0.4.1
